test(label): migrate color
