+++ core/src/components/label/test/color/label.e2e.ts-snapshots/item-color-inherit-ios-ltr-Mobile-Chrome-linux.png
