+++ core/src/components/label/test/color/label.e2e.ts-snapshots/item-color-inherit-md-ltr-Mobile-Firefox-linux.png
