+++ core/src/components/label/test/color/label.e2e.ts-snapshots/label-color-contrast-ios-ltr-Mobile-Firefox-linux.png
