+++ core/src/components/label/test/color/label.e2e.ts-snapshots/label-color-contrast-md-ltr-Mobile-Chrome-linux.png
