+++ core/src/components/label/test/color/label.e2e.ts-snapshots/label-color-ios-ltr-Mobile-Chrome-linux.png
