+++ core/src/components/label/test/color/label.e2e.ts-snapshots/label-color-md-ltr-Mobile-Firefox-linux.png
