+++ core/src/components/label/test/color/label.e2e.ts-snapshots/label-color-override-ios-ltr-Mobile-Chrome-linux.png
